backend/requirements.txt
fastapi
uvicorn[standard]
redis
requests
google-genai
better-profanity
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
websockets
python-dotenv
